+++ apps/mobile/src/screens/measurement-flow-screen/components/flow-nodes/question-node/question-types/open-ended-question.tsx
@@ -4,7 +4,7 @@ import React, { useEffect, useState } from "react";
 import { View, Text, TextInput, TouchableOpacity } from "react-native";
 import { useTheme } from "~/hooks/use-theme";
 
-import { QuestionContent } from "../../../types";
+import { QuestionContent } from "../../../../types";
 
 interface OpenEndedQuestionProps {
   content: QuestionContent;
